Add routing and layout

+++ WebApi/Services/UserService.cs
@@ -11,10 +11,10 @@ namespace WebApi.Services
     {
         User Authenticate(string email, string password);
         IEnumerable<User> GetAll();
-        User GetById(int id);
+        User GetById(long id);
         User Create(User user, string password);
         void Update(User user, string password = null);
-        void Delete(int id);
+        void Delete(long id);
     }
  
     public class UserService : IUserService
@@ -50,7 +50,7 @@ namespace WebApi.Services
             return _context.Users;
         }
  
-        public User GetById(int id)
+        public User GetById(long id)
         {
             return _context.Users.Find(id);
         }
@@ -97,7 +97,7 @@ namespace WebApi.Services
             _context.SaveChanges();
         }
  
-        public void Delete(int id)
+        public void Delete(long id)
         {
             var user = _context.Users.Find(id);
             if (user != null)
